Allow for more retries and add a summary reporter
