+++ buildutils/src/ensure-repo.ts
@@ -39,7 +39,8 @@ const UNUSED: Dict<string[]> = {
     'identity-obj-proxy',
     'jest-raw-loader',
     'markdown-loader-jest',
-    'jest-junit'
+    'jest-junit',
+    'jest-summary-reporter'
   ],
   '@jupyterlab/test-csvviewer': ['csv-spectrum'],
   '@jupyterlab/vega5-extension': ['vega', 'vega-lite'],
